agent/requirements.txt
ollama>=0.3.0
redis>=5.0.0
python-dotenv>=1.0.0
